+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Turn wifi back on after 5 seconds
 * Update AGP, Gradle, compileSdk, targetSdk, OkHttp
 * Fix 1-in-10 bug [#6](https://github.com/tir38/android-network-monkey/issues/6)
+* Only convert 200 to 404 for idempotent methods (Fixes [#2](https://github.com/tir38/android-network-monkey/issues/6))
 
 ## 1.0.2
 * Remove dependency on support lib