+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Tone down Log.e to Log.w
 * Turn wifi back on after 5 seconds
 * Update AGP, Gradle, compileSdk, targetSdk, OkHttp
+* Fix 1-in-10 bug [#6](https://github.com/tir38/android-network-monkey/issues/6)
 
 ## 1.0.2
 * Remove dependency on support lib