+++ CHANGELOG.md
@@ -1,8 +1,10 @@
-## 1.0.3 (pending)
+## 1.1.0
 * Clarify README
 * Tone down Log.e to Log.w
-* Turn wifi back on after 5 seconds
+* Turn wifi back on after 5 seconds (#4)
 * Update AGP, Gradle, compileSdk, targetSdk, OkHttp
+* Fixes "not actually 1 in 10" (#6)
+* Fixes "monkey before sending request" (#2)
 
 ## 1.0.2
 * Remove dependency on support lib